requirements.txt
wheel
django
django-allauth
django-auth-utils
django-bootstrap-datepicker-plus
django-heroku
django-bootstrap-form
django-bootstrap4
django-crispy-forms
sendgrid
pillow
boto3
django-storages
psycopg2
dj-database-url
boto3
gunicorn
Flask
asgiref
sqlparse
pytz
click
Werkzeug
Jinja2
itsdangerous
MarkupSafe
filelock
regex
bootstrap4
